pyproject.toml
[project]
name = "my_project"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "uvicorn>=0.42.0",
]
